+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Generated OpenAPI now describes whether each endpoint is forwarded (#3935).
+
 ## [2.0.2]
 
 - Avoids potential segfault when nodes receive a `SIGPIPE` interrupt from the OS (#3953).